Rename simpleAlsa module to alsa
